// src/models/User.ts
export interface User {
    address: string;
    birthdate: string;
    birthplace: string;
    city: string;
    email: string;
    fiscalCode: string;
    gender: string;
    id?: string;
    lastname: string;
    name: string;
    password: string;
    phoneNumber: number;
    piva: number;
    postalCode: number;
    role: string;
    updated: Date;
}